Reverting the version to 0.8.0-SNAPSHOT for default branch
The patch version, 0.8.1-SNAPSHOT, will be used by the release/0.8 branch after tagging the first release. It will refer to patches
on 0.8 that are in development. Version 0.8.0-SNAPSHOT is used by the default branch (master) during development of features for the
upcoming 0.8 release.

This should not be incremented by MRs. Instead, once merged, create a release candidate version on a tag commit that is separate
from the development line.